+++ UpgradeWorld/Helper.cs
@@ -21,8 +21,7 @@ public static class Helper
     if (zdo == null || !zdo.IsValid()) return;
     if (Player.m_localPlayer && Player.m_localPlayer.GetZDOID() == zdo.m_uid) return;
     if (ZNet.instance.m_peers.Any(peer => peer.m_characterID == zdo.m_uid)) return;
-    if (!zdo.IsOwner())
-      zdo.SetOwner(ZDOMan.GetSessionID());
+    zdo.SetOwner(ZDOMan.GetSessionID());
     var spawned = zdo.GetConnectionZDOID(ZDOExtraData.ConnectionType.Spawned);
     if (spawned != ZDOID.None && ZDOMan.instance.m_objectsByID.TryGetValue(spawned, out var spawnedZdo) && spawnedZdo != zdo)
       RemoveZDO(spawnedZdo);